MFC r265092, except for the ATF bits.
Fix a panic when removing an IP address from an interface, if the same address
exists on another interface.  The panic was introduced by change 264887, which
changed the fibnum parameter in the call to rtalloc1_fib() in
ifa_switch_loopback_route() from RT_DEFAULT_FIB to RT_ALL_FIBS.  The solution
is to use the interface fib in that call.  For the majority of users, that will
be equivalent to the legacy behavior.

PR:             kern/189089


git-svn-id: https://svn.freebsd.org/base/stable/9@271842 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f