MFC r272288,272289:
When setting environment variables in the atrun script, use the
"export foo=bar" form instead of "foo=bar; export foo" since the
former allows the shell to catch variable names that are not valid
shell identifiers.  This will cause /bin/sh to exit with an error
(which gets mailed to the at user) and it will not run the script.

Obtained from:	OpenBSD (r1.63 millert)
Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/stable/10@272438 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f