Support the Vodafone R215 LET USB dongle, which is apparently a rebadged E5372 with different product IDs.
Interestingly, the standard E5372 IDs (12d1:1506) are currently listed in
u3g.c and are the same as the E3131.  However, the R215/E5372 is an NCM
device and works well with cdce(4) whereas the E3131 isn't.  More work
may be needed to better identify the other device IDs.

MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@272716 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f