Use '-e' to check if the virtio backing file has already been created.
The '-f' check works fine on a regular file but not if the backing file is
a device (e.g., /dev/md0). In this case it would print a misleading but
otherwise benign message about the backing file not being present.

Submitted by:	Marcus Reid (marcus@blazingdot.com)
Discussed with:	grehan


git-svn-id: https://svn.freebsd.org/base/head@273102 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f