filedesc: cleanup setugidsafety a little
Rename it to fdsetugidsafety for consistency with other functions.

There is no need to take filedesc lock if not closing any files.

The loop has to verify each file and we are guaranteed fdtable has space
for at least 20 fds. As such there is no need to check fd_lastfile.

While here tidy up is_unsafe.


git-svn-id: https://svn.freebsd.org/base/head@273441 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f